refactor: 패키지 이름 global로 수정
