+++ src/main/java/roomescape/reservation/application/exception/ReservationException.java
@@ -1,6 +1,6 @@
 package roomescape.reservation.application.exception;
 
-import roomescape.exception.RoomEscapeException;
+import roomescape.global.RoomEscapeException;
 
 public class ReservationException extends RoomEscapeException {
     public ReservationException(String message) {
